requirements_raspberry.txt
RPi.GPIO
flask
pytest